Implement day 5
Pretty naive implementation of `react` -- it'd be better to propagate
changes in one location until the possibilities there are exhausted
before moving on, but doing that kind of contextual stuff feels a bit
trickier in Haskell, and this finishes in a minute or two -- good
enough for a bit of holiday fun.

`Min`/`Arg` and `SemiGroup` are super neat though.